wip
Half baked. The idea is that a user would manage there profile, a
profile being a singular resource, and not interact with their account
through `/users/:id/:action`. An attempt at forcing the developer to
mentally scope to the the current user (no find_by_id).

The users dashboard is an agregator (not to be confused with profile)
where objects are collected and displayed.