Added the ability to confirm email addresses.
Whenever the users email address changes, be it from github or the
preferences page, an email is sent to the user with a confirmation token
and the previous confirmation (if any) is reset. When the user follows
the confirmation link, the email address is confirmed.

Before the email address is confirmed, no daily or weekly emails will be
sent out.