Sticky Footer
In the commit 57229d03f9abded12033719355045bc90b0a99ba the footer was
made sticky by making it `position: absolute; bottom: 0`

In the commit 8976cfc968d384cba7491eaa21718ce011528dcd the sticky footer
was removed.

This commit reintroduces the stickiness of the footer. The reason for
the sticky footer is on a high resolution, the footer displays half way
up the page.