Bump mimemagic from 0.3.2 to 0.3.10
Version 0.3.2 was yanked from rubygems.org and is not available for
install anymore. Bumping to 0.3.10 which is latest patch version before
0.4.x.

Seems that the only major change is that fd.o database is not
distributed with the gem anymore, but most Linux system should already
have it preinstalled anyway. It should not cause any issues.